6 / InFileDatabase - changes InFileDatabase into InFileRepository - adjusts tests - adds small changes to FileHelper and FileHelperImpl methods

+++ src/main/java/pl/coderstrust/helpers/FileHelper.java
@@ -21,5 +21,5 @@ public interface FileHelper {
 
   boolean isEmpty() throws IOException;
 
-  void removeLine(long lineNumber) throws IOException, FileHelperException;
+  void removeLine(int lineNumber) throws IOException, FileHelperException;
 }
